+++ code/go/0chain.net/smartcontract/storagesc/models.go
@@ -642,7 +642,7 @@ type StorageAllocation struct {
 	// 00001000 - 8  - move
 	// 00010000 - 16 - copy
 	// 00100000 - 32 - rename
-	FileOptions uint8 `json:"file_options"`
+	FileOptions uint16 `json:"file_options"`
 
 	WritePool currency.Coin `json:"write_pool"`
